core: `transaction-list` module now handles rendering the list elements
